Move some reaction model tests to new test_models.py file.
